+++ jiyoontrip/codegen.ts
@@ -1,0 +1,15 @@
+import type { CodegenConfig } from "@graphql-codegen/cli";
+
+const config: CodegenConfig = {
+  overwrite: true,
+  schema: "http://main-practice.codebootcamp.co.kr/graphql",
+  documents: "src/**/*.tsx",
+  generates: {
+    "src/gql/": {
+      preset: "client",
+      plugins: [],
+    },
+  },
+};
+
+export default config;
