+++ jiyoontrip/src/app/components/boards-detail/queires.ts
@@ -1,0 +1,25 @@
+import { gql } from "@apollo/client";
+
+export const FETCH_BOARD = gql`
+  query fetchBoard($boardId: ID!) {
+    fetchBoard(boardId: $boardId) {
+      _id
+      writer
+      title
+      contents
+      likeCount
+      dislikeCount
+      images
+      boardAddress {
+        _id
+        zipcode
+        address
+        addressDetail
+        createdAt
+        updatedAt
+      }
+      createdAt
+      updatedAt
+    }
+  }
+`;
